jedire/backend/python-services/requirements.txt
# API & Web
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Cache
redis==5.0.1

# Data Processing
numpy==1.26.0
scipy==1.11.4

# Background Jobs (optional for now)
celery==5.3.4

# Auth (optional for now)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0
